+++ src/main/java/net/dv8tion/jda/internal/entities/MessageMentionsImpl.java
@@ -165,15 +165,4 @@ public class MessageMentionsImpl extends AbstractMentions
         else
             return getJDA().getRoleById(roleId);
     }
-
-    protected Emote matchEmote(Matcher m)
-    {
-        long emoteId = MiscUtil.parseSnowflake(m.group(2));
-        String name = m.group(1);
-        boolean animated = m.group(0).startsWith("<a:");
-        Emote emote = getJDA().getEmoteById(emoteId);
-        if (emote == null)
-            emote = new EmoteImpl(emoteId, jda).setName(name).setAnimated(animated);
-        return emote;
-    }
 }
